requirements.txt
-i https://pypi.org/simple
alembic==1.10.0
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
certifi==2022.12.7 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
dependency-injector==4.41.0
factory-boy==3.2.1
faker==17.6.0 ; python_version >= '3.7'
fastapi==0.92.0
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
idna==3.4
iniconfig==2.0.0 ; python_version >= '3.7'
loguru==0.6.0
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
packaging==23.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pluralizer==1.2.0
psycopg2-binary==2.9.5
pydantic==1.10.5
pytest==7.2.2
pytest-env==0.8.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
rfc3986[idna2008]==1.5.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==2.0.5.post1
starlette==0.25.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
uvicorn==0.20.0
